* #3: Refactored connection part
